pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lecturelog"
version = "0.1.0"
description = "LectureLog MVP pipeline"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "google-genai",
  "httpx",
  "fastapi",
  "uvicorn[standard]",
  "python-multipart",
  "pymupdf",
  "Pillow",
  "pydantic>=2.0",
  "pydantic-settings>=2.0",
  "aiogram>=3.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
]

[project.scripts]
lecturelog-cli = "cli.main:main"
lecturelog-bot = "bot.main:main"

[tool.setuptools.packages.find]
include = ["lecturelog*", "server*", "bot*", "cli*"]
